+++ src/jflow/db/instdata/tests/__init__.py
@@ -1,4 +1,4 @@
 
-from api import *
+#from api import *
 from forms import *
-from managers import *
+#from managers import *