+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 
 jFlow 0.4.a2
 ===========================
+ * Spin-off of django-extracontent__ from ``jflow.db.instdata``.
  * Added the ``libs`` directory containing external libraries which are not regularly updated.
  * External libraries are added to the ``PYTHONPATH``, if needed, when importing ``jflow``.
  * Portfolio views are managed by python-stdnet__ ORM.
@@ -21,4 +22,5 @@ jFlow 0.3 -  (2010 March 15)
  * Added TabularInline BondIssuer in BondClass admin
 
 
+__ http://github.com/lsbardel/django-extracontent
 __ http://packages.python.org/python-stdnet/